# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "TG Summarizer", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/rag/status": {
      "get": {
        "tags": ["rag"],
        "summary": "Rag Status",
        "operationId": "rag-rag_status",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Rag-Rag Status"
                }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/rag/embed": {
      "post": {
        "tags": ["rag"],
        "summary": "Rag Embed",
        "operationId": "rag-rag_embed",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RagEmbedRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Rag-Rag Embed"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/rag/search": {
      "post": {
        "tags": ["rag"],
        "summary": "Rag Search",
        "operationId": "rag-rag_search",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RagSearchRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Rag-Rag Search"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RagEmbedRequest": {
        "properties": {
          "limit": { "type": "integer", "title": "Limit", "default": 100 }
        },
        "type": "object",
        "title": "RagEmbedRequest"
      },
      "RagSearchRequest": {
        "properties": {
          "query": { "type": "string", "title": "Query" },
          "channels": {
            "anyOf": [
              { "items": { "type": "string" }, "type": "array" },
              { "type": "null" }
            ],
            "title": "Channels"
          },
          "startDate": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Startdate"
          },
          "endDate": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Enddate"
          },
          "limit": { "type": "integer", "title": "Limit", "default": 20 },
          "scanLimit": {
            "type": "integer",
            "title": "Scanlimit",
            "default": 5000
          }
        },
        "type": "object",
        "required": ["query"],
        "title": "RagSearchRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": { "scopes": {}, "tokenUrl": "/api/v1/login/access-token" }
        }
      }
    }
  }
}